WebSocket: Add an option to mask all outgoing frames
A callback module can send a frame (fragmented or not) using the #ws_frame{}
record. If the #ws_frame.masking_key field is defined (a 4 bytes binary)
then the payload will be masked.
But, to simplify callback modules implementation, when a websocket process
is started, the parameter "mask_frames" can be set to true. Then, all
outgoing frames will be masked. This parameter is set to false by default.