WebSocket: Add an option to set a default timeout
When a websocket process is started, the parameter "timeout" can be
defined (in milliseconds). It will be used as the default timeout.
But, all callback functions can override this value, only for the next
loop.

It is a good idea to define a default timeout to let callback module a
chance to close inactive connections.